UrlInfoController split in two controllers: SystemInfoController path of SystemInfo is now /systeminfo
