+++ bangladesh-green/src/main/java/urlshortener/bangladeshgreen/web/UrlInfoController.java
@@ -135,165 +135,6 @@ public class UrlInfoController {
         }
     }
 
-    /**
-     * Returns array of cpu or ram series or average and total clicks by day and total clicks ever if day is null
-     * also total clicks by hour at one day
-    */
-    @RequestMapping(value = "/infoday", method = RequestMethod.GET , produces ="application/json")
-    public Object locationJson(@RequestParam(value="type", required=true) String type,
-                               @RequestParam(value="day", required=false) Date day,
-                               @RequestParam(value="series", required=false) String series,
-                               HttpServletResponse response, HttpServletRequest request) {
-
-
-        //Get authentication information
-        final Claims claims = (Claims) request.getAttribute("claims");
-        String loggedRoles = (String) claims.get("roles");
-
-
-        //If global and not admin -> Forbidden
-        if(!loggedRoles.equalsIgnoreCase("admin")){
-            //Not authorized
-            ErrorResponse errorResponse = new ErrorResponse("Permission denied");
-            return new ResponseEntity<>(errorResponse, HttpStatus.FORBIDDEN);
-        }
-
-
-        List <Usage> list = null;
-
-        //CPU or RAM average in day
-        if ((type.compareTo("cpu") == 0 || type.compareTo("ram") == 0) && series.compareTo("average")==0 && day!=null){
-            double average = listCPURamAverage(day,type);
-            logger.info("(/infoday) - ("+ type +") Ok request - average: " + average);
-            SuccessResponse success = new SuccessResponse(average);
-            response.setStatus(HttpStatus.OK.value());
-            return new ResponseEntity<>(success, HttpStatus.OK);
-
-
-        }
-        //CPU or RAM series for a day. Interval: 30 seconds
-        else if ((type.compareTo("cpu") == 0 || type.compareTo("ram") == 0) && series.compareTo("series")==0 && day!=null){
-            list = listCPURam(day,type);
-            logger.info("(/infoday) - ("+ type +") Ok request - list size: " + list.size());
-            SuccessResponse success = new SuccessResponse(list);
-            response.setStatus(HttpStatus.OK.value());
-            return new ResponseEntity<>(success, HttpStatus.OK);
-
-        }
-        //Clicks aggregation, by day or total
-        else if (type.compareTo("clicks")==0){
-            //If day is null, return total clicks. Else, return clicks at the given day.
-            long total = listclicks(day);
-            logger.info("(/infoday) - (clicks) Ok request - total: " + total);
-            SuccessResponse success = new SuccessResponse(total);
-            response.setStatus(HttpStatus.OK.value());
-            return new ResponseEntity<>(success, HttpStatus.OK);
-
-        }
-        //Click list by hour for a given day
-        else if (type.compareTo("clicksadds")==0 && day!=null){
-            int hours [] = listclickshourAgregation(day);
-            SuccessResponse success = new SuccessResponse(hours);
-            response.setStatus(HttpStatus.OK.value());
-            return new ResponseEntity<>(success, HttpStatus.OK);
-        }
-        else {
-            //Bad request
-            logger.info("(/info) Bad request");
-            ErrorResponse error = new ErrorResponse("Bad request");
-            return new ResponseEntity<>(error, HttpStatus.BAD_REQUEST);
-        }
-    }
-
-    /**
-     * Returns average usage of CPU or RAM
-     */
-    private double listCPURamAverage(Date day, String type){
-        List <Usage> list = listCPURam(day,type);
-        double average = 0.0;
-        for (Usage a : list) {
-            average = average + a.getUsage();
-        }
-        return (average / list.size());
-    }
-
-    /**
-     * Returns array with clicks by hours [00-23] at a given day
-     */
-    private int [] listclickshourAgregation(Date day) {
-        Calendar calendar = Calendar.getInstance();
-        calendar.setTime(day);
-        int hours[] = new int [24];
-
-        for (int i = 0; i < 24; i++){
-            long previous_time = calendar.getTimeInMillis();
-            calendar.add(Calendar.HOUR, 1);
-            long after_time = calendar.getTimeInMillis();
-            List<Click> list = clickRepository.findByDateBetween(new Date(previous_time),new Date(after_time));
-            hours[i] = list.size();
-        }
-        return hours;
-
-    }
-
-
-    /**
-     * Returns total number of clicks one day
-     */
-    private long listclicks(Date day) {
-        List<Click> list = clickRepository.findAll();
-        long total = 0;
-        for (Click a : list) {
-            if (day == null){
-                total++;
-            }else{
-                SimpleDateFormat fmt = new SimpleDateFormat("yyyyMMdd");
-                boolean sameDay = fmt.format(a.getDate()).equals(fmt.format(day));
-                if(sameDay){
-                    total++;
-                }
-            }
-
-        }
-        return  total;
-    }
-
-    /**
-     * Returns series(time, usage) of cpu or ram at day
-     */
-    private List<Usage> listCPURam(Date day, String type) {
-        ArrayList<Usage> listt;
-        if (type.compareTo("cpu")==0){
-            List<UsageCpu> list;
-            list = cpuRepository.findAll();
-            listt = new ArrayList<>();
-            for (UsageCpu a : list) {
-                Date dateUsage = new Date (a.getTime());
-                SimpleDateFormat fmt = new SimpleDateFormat("yyyyMMdd");
-                boolean sameDay = fmt.format(dateUsage).equals(fmt.format(day));
-                if(sameDay){
-                    listt.add(a);
-                }
-            }
-        } else {
-            List<UsageRam> list;
-            list = ramRepository.findAll();
-            listt = new ArrayList<>();
-            for (UsageRam a : list) {
-                Date dateUsage = new Date (a.getTime());
-                SimpleDateFormat fmt = new SimpleDateFormat("yyyyMMdd");
-                boolean sameDay = fmt.format(dateUsage).equals(fmt.format(day));
-                if(sameDay){
-                    listt.add(a);
-                }
-            }
-        }
-        return listt;
-    }
-
-
-
-
 
 
     /**
